pdc: Fix compiler warning
When building with clang it warns:

zephyr/subsys/pd_controller/pdc_console.c:222:52: error: result of
comparison of constant 255 with expression of type 'char' is always true
[-Werror,-Wtautological-constant-out-of-range-compare]
  222 |                           pdc_info.project_name[0] != 0xFF;
      |                           ~~~~~~~~~~~~~~~~~~~~~~~~ ^  ~~~~

When char is signed, it only has a range of -128 to 127. Cast it to a
uint8_t so the comparison to 255 is valid.

BUG=b:390722919
TEST=CQ

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6206776
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Tested-by: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Tom Hughes <tomhughes@chromium.org>

+++ zephyr/subsys/pd_controller/pdc_console.c
@@ -219,7 +219,7 @@ static int cmd_pdc_get_info(const struct shell *sh, size_t argc, char **argv)
 
 	/* Check if the FW project name is set. */
 	bool has_proj_name = pdc_info.project_name[0] != '\0' &&
-			     pdc_info.project_name[0] != 0xFF;
+			     (uint8_t)pdc_info.project_name[0] != 0xFF;
 
 	shell_fprintf(sh, SHELL_INFO,
 		      "Live: %d\n"
